refactor(export): split step-serialization.ts on the argument-parser and file-assembly seams
step-serialization.ts was 678 lines, past the ~400-line module guideline, and
was carrying three unrelated jobs. Two of them move out; both moves are
byte-identical code with no behaviour change and no change to the package's
published surface (none of the moved symbols is exported from index.ts).

step-argument-parser.ts takes splitTopLevelArgs, replaceStepArgument and
splitTopLevelStepArguments — the one text layer that runs the other way from
the rest of the export path, reading a record's slots back out of a line and
writing one back by index. They share one set of rules (quote state,
doubled-quote escapes, paren depth, what counts as a slot), and those rules are
what a caller needs to find when a rewritten line comes out wrong; the two
recently hardened functions now sit together instead of a hundred lines apart.

step-file-assembly.ts takes assembleStepBytes and assembleStepBlob, which
serialize nothing: they join a finished header and finished entity lines into
the delivered artifact, and their contract is byte-identity with each other.

The test file splits the same way, each block moving with the code it pins.
Test count is unchanged (648 passing / 30 skipped); five guard mutations from
the moved code each kill exactly the same tests before and after the move.

Refs #2475
