+++ packages/export/src/step-serialization.test.ts
@@ -5,8 +5,6 @@
 import { describe, expect, it } from 'vitest';
 import { PropertyValueType } from '@ifc-lite/data';
 import {
-  assembleStepBlob,
-  assembleStepBytes,
   serializePropertyValue,
   serializeAttributeValue,
   serializeStepValue,
@@ -14,8 +12,6 @@ import {
   resolveExpressBase,
   tokenIsRealLiteral,
   toStepReal,
-  replaceStepArgument,
-  splitTopLevelStepArguments,
 } from './step-serialization.js';
 import { toStepRealScaled } from './unit-normalize.js';
 
@@ -196,270 +192,5 @@ describe('toStepRealScaled', () => {
     for (const v of [Number.MAX_VALUE, Number.MIN_VALUE, -1.5e-300, 1e-7, 123.456]) {
       expect(toStepRealScaled(v)).toMatch(STEP_REAL_RE);
     }
-  });
-});
-
-/**
- * Reference implementation of the OLD (pre-rewrite) `assembleStepBytes`:
- * single-pass `encoder.encode()` per entity, keeping every encoded chunk
- * alive in a persistent `Uint8Array[]` until the final copy. Kept here
- * (rather than trusting a snapshot) so the byte-identity test fails loudly
- * if the new two-pass `encodeInto` assembler ever drifts from it, on a
- * UTF-8 corpus that specifically exercises multi-byte characters.
- */
-function assembleStepBytesReference(header: string, entities: string[]): Uint8Array {
-  const encoder = new TextEncoder();
-  const headBytes = encoder.encode(`${header}DATA;\n`);
-  const tailBytes = encoder.encode('ENDSEC;\nEND-ISO-10303-21;\n');
-  const newline = encoder.encode('\n');
-
-  let totalSize = headBytes.byteLength + tailBytes.byteLength;
-  const entityBytes: Uint8Array[] = new Array(entities.length);
-  for (let i = 0; i < entities.length; i++) {
-    entityBytes[i] = encoder.encode(entities[i]);
-    totalSize += entityBytes[i].byteLength + newline.byteLength;
-  }
-
-  const result = new Uint8Array(totalSize);
-  let offset = 0;
-  result.set(headBytes, offset);
-  offset += headBytes.byteLength;
-  for (let i = 0; i < entityBytes.length; i++) {
-    result.set(entityBytes[i], offset);
-    offset += entityBytes[i].byteLength;
-    result.set(newline, offset);
-    offset += newline.byteLength;
-  }
-  result.set(tailBytes, offset);
-  return result;
-}
-
-const HEADER = "ISO-10303-21;\nHEADER;\nFILE_DESCRIPTION((''),'2;1');\nENDSEC;\n";
-
-/** UTF-8 corpus: ASCII, Latin-1 accents, 2/3-byte BMP chars, and a 4-byte
- * surrogate-pair emoji, all inside STEP entity strings (the realistic case:
- * IFCLABEL/IFCTEXT attribute values carry user text). */
-const UTF8_ENTITIES = [
-  "#1=IFCWALL('0000000000000000000001',$,'Plain ASCII wall',$,$,$,$,$,$);",
-  "#2=IFCLABEL('Wand mit Umlauten: äöüÄÖÜß und Zeichen: café, naïve');",
-  "#3=IFCTEXT('日本語のテキスト and 中文文本 mixed with ASCII');",
-  "#4=IFCLABEL('Emoji stress test: 🏗️🏢🧱🪟 and combining marks: é');",
-  "#5=IFCLABEL('');", // empty string entity
-  `#6=IFCTEXT('${'x'.repeat(5000)}${'ü'.repeat(2000)}${'文'.repeat(1000)}');`, // forces scratch-buffer growth
-];
-
-describe('assembleStepBytes', () => {
-  it('is byte-identical to the pre-rewrite single-pass reference on ASCII-only entities', () => {
-    const entities = [
-      "#1=IFCPROJECT('0000000000000000000001',$,'Project',$,$,$,$,$,$);",
-      "#2=IFCWALL('0000000000000000000002',$,'Wall 1',$,$,$,$,$,$);",
-    ];
-    const expected = assembleStepBytesReference(HEADER, entities);
-    const actual = assembleStepBytes(HEADER, entities);
-    expect(actual).toEqual(expected);
-  });
-
-  it('is byte-identical to the pre-rewrite reference on a multi-byte UTF-8 corpus', () => {
-    const expected = assembleStepBytesReference(HEADER, UTF8_ENTITIES);
-    const actual = assembleStepBytes(HEADER, UTF8_ENTITIES);
-    expect(actual.length).toBe(expected.length);
-    expect(actual).toEqual(expected);
-  });
-
-  it('handles zero entities', () => {
-    const expected = assembleStepBytesReference(HEADER, []);
-    const actual = assembleStepBytes(HEADER, []);
-    expect(actual).toEqual(expected);
-  });
-
-  it('round-trips through TextDecoder back to the original entity text', () => {
-    const bytes = assembleStepBytes(HEADER, UTF8_ENTITIES);
-    const text = new TextDecoder('utf-8').decode(bytes);
-    for (const entity of UTF8_ENTITIES) {
-      expect(text).toContain(entity);
-    }
-  });
-});
-
-describe('assembleStepBlob', () => {
-  it('has byte content identical to assembleStepBytes on a multi-byte UTF-8 corpus', async () => {
-    const blob = assembleStepBlob(HEADER, UTF8_ENTITIES);
-    const blobBytes = new Uint8Array(await blob.arrayBuffer());
-    const bytes = assembleStepBytes(HEADER, UTF8_ENTITIES);
-    expect(blobBytes).toEqual(bytes);
-  });
-
-  it('has byte content identical to assembleStepBytes on ASCII-only entities', async () => {
-    const entities = [
-      "#1=IFCPROJECT('0000000000000000000001',$,'Project',$,$,$,$,$,$);",
-      "#2=IFCWALL('0000000000000000000002',$,'Wall 1',$,$,$,$,$,$);",
-    ];
-    const blob = assembleStepBlob(HEADER, entities);
-    const blobBytes = new Uint8Array(await blob.arrayBuffer());
-    const bytes = assembleStepBytes(HEADER, entities);
-    expect(blobBytes).toEqual(bytes);
-  });
-
-  it('handles zero entities identically to assembleStepBytes', async () => {
-    const blob = assembleStepBlob(HEADER, []);
-    const blobBytes = new Uint8Array(await blob.arrayBuffer());
-    const bytes = assembleStepBytes(HEADER, []);
-    expect(blobBytes).toEqual(bytes);
-  });
-
-  it('reports the combined byte size via blob.size', async () => {
-    const blob = assembleStepBlob(HEADER, UTF8_ENTITIES);
-    const bytes = assembleStepBytes(HEADER, UTF8_ENTITIES);
-    expect(blob.size).toBe(bytes.byteLength);
-  });
-});
-
-describe('replaceStepArgument slot validation', () => {
-  const LINE = "#5=IFCWALLTYPE('0OSuGGYUFyIf0LtE29OSuT',$,'WT1',$,$,(#30),$,$,$,.STANDARD.);";
-
-  it('replaces the requested slot and leaves every other token byte-identical', () => {
-    const out = replaceStepArgument(LINE, 5, '(#33)');
-    expect(out).toBe("#5=IFCWALLTYPE('0OSuGGYUFyIf0LtE29OSuT',$,'WT1',$,$,(#33),$,$,$,.STANDARD.);");
-  });
-
-  it('returns null for a slot past the end rather than a silently unchanged line', () => {
-    expect(replaceStepArgument(LINE, 99, '(#33)')).toBeNull();
-  });
-
-  // A negative or fractional index would assign a NAMED PROPERTY on the array
-  // instead of an element; `join` skips it, so the function would hand back the
-  // line unchanged but NON-NULL. Callers read non-null as "the replacement
-  // happened" -- `rewriteTypeOwnedPsetLine` turns it into `repointed: true` --
-  // so an unchanged non-null is a silent false success, the same shape as the
-  // drop this module's caller was fixed for. Unreachable through the exporter
-  // today (the only slot is a constant), pinned because the contract is public.
-  for (const slot of [-1, -5, 1.5, Number.NaN]) {
-    it(`returns null for the invalid slot ${String(slot)}`, () => {
-      expect(replaceStepArgument(LINE, slot, '(#33)')).toBeNull();
-    });
-  }
-});
-
-/**
- * github.com/LTplus-AG/ifc-lite/issues/2470, second half: the helper's failure
- * SIGNALLING, one level below the null contract above.
- *
- * `splitTopLevelStepArguments` tracked quote state and paren depth to find the
- * top-level commas and then ignored where that scan ended up. Text that never
- * left a string, or never closed a nested list, still produced parts — parts
- * whose boundaries are wherever the scanner stopped rather than the record's
- * slots. `replaceStepArgument`'s regex pins only the two ENDS of the record, so
- * such a line reaches the split, gets a slot written by index, and comes back
- * NON-NULL: a success it did not achieve, and a corrupted line where #2469 had
- * a dropped one.
- *
- * Each malformed case below returned a string before the fix — the mutation
- * check for this block is to delete one rejection and watch its case go from
- * `null` to a plausible-looking rewritten line. The VALID block underneath is
- * the bounding control: rejecting everything would pass the block above alone,
- * and would take the type-object repoint down with it (every real IfcWallType
- * line goes through here).
- */
-describe('replaceStepArgument rejects an argument list it could not scan', () => {
-  it('returns null for an unterminated quoted string', () => {
-    // The quote before `WT1` never closes, so everything after it is one
-    // "string" and the remaining commas are invisible to the scan.
-    expect(
-      replaceStepArgument("#5=IFCWALLTYPE('0OSuGGYU',$,'WT1,$,$,(#30),$);", 1, "'X'"),
-    ).toBeNull();
-  });
-
-  it('returns null for an unbalanced nested list', () => {
-    // `(#30` never closes: the scan ends at depth 1 having swallowed every
-    // comma after it.
-    expect(
-      replaceStepArgument("#5=IFCWALLTYPE('0OSuGGYU',$,'WT1',$,$,(#30,$,$,$,.STANDARD.);", 1, "'X'"),
-    ).toBeNull();
-  });
-
-  it('returns null for a stray closing paren the scan recovers from', () => {
-    // Depth goes NEGATIVE at the paren after `'a'` and back to zero at the one
-    // before `'b'`, so the FINAL state is balanced and a final-state check alone
-    // calls this well-formed. It is not: every comma while depth was negative
-    // was swallowed, so the parts that come out are `["'a'),('b'", '$', '$',
-    // '$', '$']` and writing slot 1 lands on an argument the record does not
-    // have — the corrupted-output case, returned as a success.
-    expect(replaceStepArgument("#7=IFCFOO('a'),('b',$,$,$,$);", 1, "'X'")).toBeNull();
-  });
-
-  it('writes the slot on a line with an EMPTY one rather than refusing it', () => {
-    // Invalid STEP, and deliberately still accepted: an empty argument is ONE
-    // part, exactly as the entity parser counts it, so slot 5 is still slot 5.
-    // Refusing it is what would do damage — the parser resolves
-    // `HasPropertySets` on such a line, so by the time the repoint runs the
-    // export has already withheld the property set's own lines, and a refused
-    // repoint leaves the record pointing at an entity that is no longer in the
-    // file. The exporter-level case is `a line the parser accepted keeps its
-    // slot 5 in step with the psets it dropped` in `slot5-caller-audit.test.ts`.
-    expect(replaceStepArgument("#5=IFCWALLTYPE('0OSuGGYU',,'WT1',$,$,(#30));", 5, '(#33)'))
-      .toBe("#5=IFCWALLTYPE('0OSuGGYU',,'WT1',$,$,(#33));");
-    expect(replaceStepArgument("#5=IFCWALLTYPE('0OSuGGYU',$,'WT1',$,$,(#30),);", 5, '(#33)'))
-      .toBe("#5=IFCWALLTYPE('0OSuGGYU',$,'WT1',$,$,(#33),);");
-  });
-
-  it('returns null for a record with no arguments at all', () => {
-    // Not an empty SLOT — a record that has no slots, so slot 0 is still past
-    // the end and the answer is the same null the bounds check gives.
-    expect(replaceStepArgument('#5=IFCWALLTYPE();', 0, '(#33)')).toBeNull();
-  });
-});
-
-describe('splitTopLevelStepArguments contract', () => {
-  it('splits a well-formed list and keeps each token verbatim', () => {
-    expect(splitTopLevelStepArguments("'a',$,(#1,#2),.T.")).toEqual(["'a'", '$', '(#1,#2)', '.T.']);
-  });
-
-  it('reads an EMPTY argument list as no arguments, not as a malformed one', () => {
-    // `#1=IFCFOO();` is a record with no slots — different from `(,)`, which is
-    // a record with a slot nothing was written into. Callers that ask for a
-    // slot get their answer from the bounds check instead.
-    expect(splitTopLevelStepArguments('')).toEqual([]);
-    expect(splitTopLevelStepArguments('   ')).toEqual([]);
-  });
-
-  it('returns null, not partial parts, for each way the scan can end badly', () => {
-    expect(splitTopLevelStepArguments("'a',$,'unterminated")).toBeNull();
-    expect(splitTopLevelStepArguments("'a',(#1,#2")).toBeNull();
-    expect(splitTopLevelStepArguments("'a'),('b',$")).toBeNull();
-  });
-
-  it('keeps an empty slot as a slot, so every index after it still lines up', () => {
-    // The only malformity that does NOT shift the parts: one empty argument is
-    // one part, which is how the entity parser reads it too.
-    expect(splitTopLevelStepArguments("'a',,$")).toEqual(["'a'", '', '$']);
-    expect(splitTopLevelStepArguments("'a',$,")).toEqual(["'a'", '$', '']);
-  });
-});
-
-describe('replaceStepArgument still accepts every well-formed list', () => {
-  it('keeps a comma inside a quoted string out of the split', () => {
-    const line = "#5=IFCWALLTYPE('0OSuGGYU',$,'WT1, exterior',$,$,(#30),$,$,$,.STANDARD.);";
-    expect(replaceStepArgument(line, 5, '(#33)')).toBe(
-      "#5=IFCWALLTYPE('0OSuGGYU',$,'WT1, exterior',$,$,(#33),$,$,$,.STANDARD.);",
-    );
-  });
-
-  it('keeps a doubled-quote escape and the parens inside a string intact', () => {
-    const line = "#5=IFCWALLTYPE('0OSuGGYU',$,'O''Brien (west),$',$,$,(#30),$,$,$,.STANDARD.);";
-    expect(replaceStepArgument(line, 5, '(#33)')).toBe(
-      "#5=IFCWALLTYPE('0OSuGGYU',$,'O''Brien (west),$',$,$,(#33),$,$,$,.STANDARD.);",
-    );
-  });
-
-  it('keeps a nested list argument intact', () => {
-    const line = '#7=IFCFOO((1.,2.,3.),$,(#1,#2),$,$,(#30));';
-    expect(replaceStepArgument(line, 5, '(#33)')).toBe('#7=IFCFOO((1.,2.,3.),$,(#1,#2),$,$,(#33));');
-  });
-
-  it('keeps a multi-line record intact', () => {
-    const line = "#5=IFCWALLTYPE('0OSuGGYU',\n$,\n'WT1',$,$,(#30),$,$,$,.STANDARD.);";
-    expect(replaceStepArgument(line, 5, '(#33)')).toBe(
-      "#5=IFCWALLTYPE('0OSuGGYU',\n$,\n'WT1',$,$,(#33),$,$,$,.STANDARD.);",
-    );
   });
 });
